Organize frontend components - Add pages directory for page components - Add layouts directory for layout components - Organize remaining components into features directory, with hooks, components, context, services directories of their own * status-update and ReportButton will be organized in upcoming update
