+++ frontend/src/features/posts/components/SearchBar.jsx
