+++ frontend/src/features/profiles/components/ProfileSection.jsx
