+++ frontend/src/features/votes/components/VoteDisplaySkeleton.jsx
