+++ frontend/src/hooks/index.js
@@ -1,4 +1,3 @@
 import { useVotedPosts } from "./useVotedPosts";
-import { useVote } from "./useVote";
 import { useSessionStorage } from "./useSessionStorage";
-export { useVotedPosts, useVote, useSessionStorage };
+export { useVotedPosts, useSessionStorage };
