+++ frontend/src/pages/Logout.jsx
@@ -1,0 +1,7 @@
+import LoggedOut from "../features/authentication/components/LoggedOut";
+
+const Logout = () => {
+  return <LoggedOut />;
+};
+
+export default Logout;
