+++ frontend/src/pages/NotPermitted.jsx
